refactor(assistant): make the persistence-lifecycle seam a memory-plugin concept
The MemoryPersistenceHooks seam is memory-domain end to end — every event
and query on it exists for the memory feature — but it lived in
persistence/ and its registration lived in the defaults barrel,
masquerading as host/default logic. Move both into the memory plugin,
until these events are exposed through the first-class hooks system:

- plugins/defaults/memory/persistence-lifecycle-seam.ts — the hook
  contract, no-op default, and registered-handler slot. Deliberately a
  leaf module (type-only imports): persistence call sites import it while
  the handler implementation transitively imports persistence, so any
  runtime dependency here would close an import cycle. Verified the
  branch's cycle set is identical to main's.
- plugins/defaults/memory/persistence-hooks-registration.ts — the
  disabled-state guard and registerDefaultPluginPersistenceHooks, moved
  out of the defaults barrel. The standalone memory jobs worker
  (jobs/worker.ts) and plugin bootstrap now call it from here.
- persistence/memory-lifecycle-hooks.ts is deleted; its tests move to
  the plugin as persistence-lifecycle-seam.test.ts.

The two persistence call sites (conversation-crud, jobs-worker) now
import the seam from the plugin — pinned in the persistence-layering
guard's allowlist with the decoupling plan: the memory jobs worker moves
into the plugin (retiring the jobs-worker entry) and per-event lifecycle
notifications migrate to the hooks system (retiring the
conversation-crud entry).
