+++ assistant/src/__tests__/persistence-layering-guard.test.ts
@@ -37,16 +37,30 @@ const MEMORY_DIR = join(ASSISTANT_SRC, "plugins", "defaults", "memory");
  * TECH DEBT — residual `persistence/` → `memory/` feature back-imports.
  *
  * A genuine coupling to the memory *feature* that violates the one-way
- * memory → persistence direction would be pinned here so the guard fails the
- * moment a NEW back-import is introduced. It is currently EMPTY: persistence
- * reaches the memory feature only through the `memory-lifecycle-hooks` seam
- * (persistence calls registered handlers; memory never gets imported), never by
- * importing memory internals. Do not add an entry without a decoupling plan.
+ * memory → persistence direction is pinned here so the guard fails the moment
+ * a NEW back-import is introduced. Do not add an entry without a decoupling
+ * plan.
+ *
+ * Current entries: the memory plugin owns its persistence-lifecycle seam
+ * (`persistence-lifecycle-seam.ts` — every event and query on it is
+ * memory-domain), and these persistence call sites invoke the seam's
+ * registered handlers through `getMemoryPersistenceHooks()` (no-op when
+ * memory is not present). Decoupling plan: the memory jobs worker moves into
+ * the memory plugin (retiring the jobs-worker entry), and the per-event
+ * lifecycle notifications migrate to the first-class `hooks` system (retiring
+ * the conversation-crud entry).
  *
  * Keyed by the importing persistence file (relative to repo root); the value
  * is the set of allowed `memory/<specifier>` module paths it may import.
  */
-const PERSISTENCE_TO_MEMORY_ALLOWLIST: Record<string, ReadonlySet<string>> = {};
+const PERSISTENCE_TO_MEMORY_ALLOWLIST: Record<string, ReadonlySet<string>> = {
+  "assistant/src/persistence/conversation-crud.ts": new Set([
+    "persistence-lifecycle-seam",
+  ]),
+  "assistant/src/persistence/jobs-worker.ts": new Set([
+    "persistence-lifecycle-seam",
+  ]),
+};
 
 /**
  * PERMANENT exception — the migration registry, NOT tech debt.
