+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -192,7 +192,6 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../persistence/job-utils.js",
     "../../../persistence/jobs-store.js",
     "../../../persistence/jobs-worker.js",
-    "../../../persistence/memory-lifecycle-hooks.js",
     "../../../persistence/message-content.js",
     "../../../persistence/raw-query.js",
     "../../../persistence/schema/index.js",
@@ -213,6 +212,7 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../util/sqlite-retry.js",
     "../../../util/strip-comment-lines.js",
     "../../../util/worker-memory.js",
+    "../../disabled-state.js",
     "../../types.js",
     "../injection-presence.js",
     "../injector-order.js",
